requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
numpy==1.24.1
pandas==1.5.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
rest-pandas==1.1.0
six==1.16.0
sqlparse==0.4.3
